refactor: extract CardCache to CardRepository
