+++ src/main/java/Application.java
@@ -1,5 +1,6 @@
 import controller.BlackjackController;
 import domain.card.Card;
+import domain.card.CardRepository;
 import domain.card.Cards;
 import domain.card.Deck;
 import domain.user.Dealer;
@@ -15,7 +16,7 @@ import domain.user.Players;
  */
 public class Application {
 	public static void main(String[] args) {
-		Deck deck = new Deck(Card.CardCache.toList());
+		Deck deck = new Deck(CardRepository.toList());
 		Dealer dealer = new Dealer();
 		BlackjackController.run(deck, dealer, Players.of("kouz,toney"));
 	}
